requirements.txt
numpy
pillow
pytest
requests
pytest-mock
faker
